knowledge-bases/03-infra/e2e-rag-using-bedrock-kb-cdk/requirements.txt
aws-cdk-lib==2.141.0
constructs>=10.0.0,<11.0.0
